Wrapper of glClearColor and corresponding glGet
Added a trivial thin wrapper of glGet to take advantage of function
overloading.  Used this in a thicker wrapper of glClearColor and
glGet(GL_COLOR_CLEAR_VALUE, ...) so that the two can be found in the same
place, rather than requiring completely different calls.  The same
infrastructure should support all analogous GL constructs.

Signed-off-by: John Bytheway <jbytheway@gmail.com>